Cairo 1.16.0

+++ Libraries/3rd/cairo/source/src/cairo-image-surface-private.h
@@ -40,6 +40,7 @@
 
 #include "cairo-surface-private.h"
 
+#include <stddef.h>
 #include <pixman.h>
 
 CAIRO_BEGIN_DECLS
@@ -71,7 +72,7 @@ struct _cairo_image_surface {
 
     int width;
     int height;
-    int stride;
+    ptrdiff_t stride;
     int depth;
 
     unsigned owns_data : 1;
